# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  username: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  username?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Activity {
  id: number;
  name: string;
  category: string;
  color: string;
  /** @nullable */
  icon?: string | null;
  targetMinutesPerDay: number;
  createdAt: string;
}

export interface ActivityInput {
  /** @minLength 1 */
  name: string;
  category: string;
  color: string;
  icon?: string;
  /** @minimum 1 */
  targetMinutesPerDay: number;
}

export interface ActivityPatch {
  /** @minLength 1 */
  name?: string;
  category?: string;
  color?: string;
  icon?: string;
  /** @minimum 1 */
  targetMinutesPerDay?: number;
}

export interface ActivityLog {
  id: number;
  activityId: number;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  logDate: string;
  createdAt: string;
}

export interface ActivityLogInput {
  /** @minimum 1 */
  durationMinutes: number;
  notes?: string;
  /** YYYY-MM-DD, defaults to today */
  logDate?: string;
}

export interface Alert {
  id: number;
  activityId: number;
  /** @nullable */
  activityName?: string | null;
  /** HH:MM */
  timeOfDay: string;
  /** 0=Sun, 1=Mon, ... 6=Sat */
  daysOfWeek: number[];
  enabled: boolean;
  message: string;
}

export interface AlertInput {
  activityId: number;
  timeOfDay: string;
  daysOfWeek: number[];
  enabled?: boolean;
  message: string;
}

export interface AlertPatch {
  timeOfDay?: string;
  daysOfWeek?: number[];
  enabled?: boolean;
  message?: string;
}

export interface Achievement {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  activityName?: string | null;
  unlockedAt: string;
}

export interface Streak {
  activityId: number;
  activityName: string;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastLoggedDate: string | null;
}

export interface DashboardSummary {
  totalActivities: number;
  totalMinutesToday: number;
  activitiesTodayCompleted: number;
  activitiesTodayTotal: number;
  overallCurrentStreak: number;
  totalAchievements: number;
  recentAchievements: Achievement[];
  todayLogs: ActivityLog[];
}

export interface DayProgress {
  date: string;
  minutesLogged: number;
  completed: boolean;
}

export interface WeeklyProgress {
  activityId: number;
  activityName: string;
  color: string;
  targetMinutesPerDay?: number;
  days: DayProgress[];
}