# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  activityId?: number | null;
  /** @nullable */
  activityName?: string | null;
  unlockedAt: string;
}