# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  /** @minLength 1 */
  name: string;
  category: string;
  color: string;
  icon?: string;
  /** @minimum 1 */
  targetMinutesPerDay: number;
}