# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  activityId: number;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  logDate: string;
  createdAt: string;
}