# lib/api-zod/src/generated/types/activityLogInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLogInput {
  /** @minimum 1 */
  durationMinutes: number;
  notes?: string;
  /** YYYY-MM-DD, defaults to today */
  logDate?: string;
}