# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  activityId: number;
  /** @nullable */
  activityName?: string | null;
  /** HH:MM */
  timeOfDay: string;
  /** 0=Sun, 1=Mon, ... 6=Sat */
  daysOfWeek: number[];
  enabled: boolean;
  message: string;
}