# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Achievement } from './achievement';
import type { ActivityLog } from './activityLog';

export interface DashboardSummary {
  totalActivities: number;
  totalMinutesToday: number;
  activitiesTodayCompleted: number;
  activitiesTodayTotal: number;
  overallCurrentStreak: number;
  totalAchievements: number;
  recentAchievements: Achievement[];
  todayLogs: ActivityLog[];
}