# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  username: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}