# lib/api-zod/src/generated/types/weeklyProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Learning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DayProgress } from './dayProgress';

export interface WeeklyProgress {
  activityId: number;
  activityName: string;
  color: string;
  targetMinutesPerDay?: number;
  days: DayProgress[];
}